Refactor into gpt_trader package
