+++ src/gpt_trader/cli/main_liveTrade.py
@@ -9,10 +9,9 @@ import json
 import tempfile
 import logging
 import sys
-from datetime import datetime, timedelta
 from pathlib import Path
 
-from scripts.common import _run_step
+from gpt_trader.cli.common import _run_step
 
 
 def _load_config(path: Path) -> dict:
@@ -28,7 +27,7 @@ async def main() -> None:
     default_cfg = (
         Path(__file__).resolve().parent
         / "config"
-        / "setting_main_backtest.json"
+        / "setting_main_liveTrade.json"
     )
     pre_parser.add_argument(
         "--config", help="Path to JSON config", default=str(default_cfg)
@@ -62,20 +61,20 @@ async def main() -> None:
     )
     parser.add_argument(
         "--send-script",
-        default=scripts_cfg.get("send", "scripts/send_api/send_to_gpt.py"),
+        default=scripts_cfg.get("send", "src/gpt_trader/send/send_to_gpt.py"),
         help="Path to GPT API script",
     )
     parser.add_argument(
         "--parse-script",
         default=scripts_cfg.get(
-            "parse", "scripts/parse_response/parse_gpt_response.py"
+            "parse", "src/gpt_trader/parse/parse_gpt_response.py"
         ),
         help="Path to response parsing script",
     )
     parser.add_argument(
         "--response",
         default=workflow.get(
-            "response", "data/back_test/signals/latest_response.txt"
+            "response", "live_trade/data/signals/latest_response.txt"
         ),
         help="Temporary file to store raw GPT response",
     )
@@ -103,17 +102,10 @@ async def main() -> None:
     send_cfg = config.get("send")
     parse_cfg = config.get("parse")
 
-    start_time = datetime.fromisoformat(config.get("start_time"))
-    end_time = datetime.fromisoformat(config.get("end_time"))
-    step = timedelta(minutes=int(config.get("loop_every_minutes", 60)))
-    signal_table = Path(config.get(
-        "signal_table", "data/back_test/signals/backtest_signals.csv"
-    ))
-
     if not args.fetch_script:
         fetch_map = {
-            "yf": "scripts/fetch/fetch_yf_data.py",
-            "mt5": "scripts/fetch/fetch_mt5_data.py",
+            "yf": "src/gpt_trader/fetch/fetch_yf_data.py",
+            "mt5": "src/gpt_trader/fetch/fetch_mt5_data.py",
         }
         args.fetch_script = fetch_map[args.fetch_type]
 
@@ -122,54 +114,41 @@ async def main() -> None:
         format="%(asctime)s [%(levelname)s] %(message)s",
     )
 
-    current = start_time
-    while current <= end_time:
-        logging.info("Backtest step at %s", current.isoformat())
-
-        step_fetch = fetch_cfg.copy() if fetch_cfg else None
-        if step_fetch is not None:
-            step_fetch["time_fetch"] = current.strftime("%Y-%m-%d %H:%M:%S")
-
-        step_parse = parse_cfg.copy() if parse_cfg else {}
-        step_parse["path_signals_csv"] = str(signal_table.parent)
-        step_parse["file_signal_report"] = signal_table.name
-
-        if not args.skip_fetch:
-            if step_fetch is not None:
-                with tempfile.NamedTemporaryFile("w", delete=False, suffix=".json") as tmp:
-                    json.dump(step_fetch, tmp)
-                try:
-                    await _run_step("fetch", Path(args.fetch_script), "--config", tmp.name)
-                finally:
-                    Path(tmp.name).unlink(missing_ok=True)
-            else:
-                await _run_step("fetch", Path(args.fetch_script))
-
-        if not args.skip_send:
-            send_args = ["--output", args.response]
-            if send_cfg:
-                with tempfile.NamedTemporaryFile("w", delete=False, suffix=".json") as tmp:
-                    json.dump(send_cfg, tmp)
-                send_args.extend(["--config", tmp.name])
-                try:
-                    await _run_step("send", Path(args.send_script), *send_args)
-                finally:
-                    Path(tmp.name).unlink(missing_ok=True)
-            else:
-                await _run_step("send", Path(args.send_script), *send_args)
-
-        if not args.skip_parse:
-            parse_args = []
+    if not args.skip_fetch:
+        if fetch_cfg:
             with tempfile.NamedTemporaryFile("w", delete=False, suffix=".json") as tmp:
-                json.dump(step_parse, tmp)
+                json.dump(fetch_cfg, tmp)
+            try:
+                await _run_step("fetch", Path(args.fetch_script), "--config", tmp.name)
+            finally:
+                Path(tmp.name).unlink(missing_ok=True)
+        else:
+            await _run_step("fetch", Path(args.fetch_script))
+    if not args.skip_send:
+        send_args = ["--output", args.response]
+        if send_cfg:
+            with tempfile.NamedTemporaryFile("w", delete=False, suffix=".json") as tmp:
+                json.dump(send_cfg, tmp)
+            send_args.extend(["--config", tmp.name])
+            try:
+                await _run_step("send", Path(args.send_script), *send_args)
+            finally:
+                Path(tmp.name).unlink(missing_ok=True)
+        else:
+            await _run_step("send", Path(args.send_script), *send_args)
+    if not args.skip_parse:
+        parse_args = []
+        if parse_cfg:
+            with tempfile.NamedTemporaryFile("w", delete=False, suffix=".json") as tmp:
+                json.dump(parse_cfg, tmp)
             parse_args.extend(["--config", tmp.name])
             parse_args.append(args.response)
             try:
                 await _run_step("parse", Path(args.parse_script), *parse_args)
             finally:
                 Path(tmp.name).unlink(missing_ok=True)
-
-        current += step
+        else:
+            await _run_step("parse", Path(args.parse_script), args.response)
 
 
 if __name__ == "__main__":
