+++ src/gpt_trader/parse/config/parse.json
