+++ src/gpt_trader/parse/parse_gpt_response.py
