+++ src/gpt_trader/send/config/gpt.example.json
