+++ src/gpt_trader/send/send_to_gpt.py
