+++ tests/test_fetch_yf_data.py
@@ -1,4 +1,3 @@
-import sys
 from pathlib import Path
 from unittest.mock import patch
 import json
@@ -8,9 +7,7 @@ import importlib
 import pandas as pd
 import pytest
 
-sys.path.insert(0, str(Path(__file__).resolve().parents[1]))
-
-from scripts.fetch.fetch_yf_data import fetch_multi_tf
+from gpt_trader.fetch.fetch_yf_data import fetch_multi_tf
 
 
 def _fake_download(
@@ -32,7 +29,7 @@ def _fake_download(
 
 def test_fetch_multi_tf_returns_dataframe() -> None:
     config = {"fetch_bars": 5, "timeframes": [{"tf": "M1", "keep": 5}]}
-    with patch("scripts.fetch.fetch_yf_data.yf.download", side_effect=_fake_download):
+    with patch("gpt_trader.fetch.fetch_yf_data.yf.download", side_effect=_fake_download):
         df = fetch_multi_tf("TEST", config, tz_shift=0)
     assert isinstance(df, pd.DataFrame)
     assert not df.empty
@@ -41,7 +38,7 @@ def test_fetch_multi_tf_returns_dataframe() -> None:
 def test_tz_shift_applied() -> None:
     """Timestamps should shift by the tz_shift value."""
     config = {"fetch_bars": 5, "timeframes": [{"tf": "M1", "keep": 5}]}
-    with patch("scripts.fetch.fetch_yf_data.yf.download", side_effect=_fake_download):
+    with patch("gpt_trader.fetch.fetch_yf_data.yf.download", side_effect=_fake_download):
         df = fetch_multi_tf("TEST", config, tz_shift=3)
 
     expected = pd.date_range("2024-01-01", periods=5, freq="min") + pd.Timedelta(
@@ -62,9 +59,9 @@ def test_main_error_on_empty_df(tmp_path, caplog) -> None:
     cfg_path.write_text(json.dumps(cfg))
 
     with patch.object(sys, "argv", ["fetch_yf_data.py", "--config", str(cfg_path)]), patch(
-        "scripts.fetch.fetch_yf_data.fetch_multi_tf", return_value=pd.DataFrame()
+        "gpt_trader.fetch.fetch_yf_data.fetch_multi_tf", return_value=pd.DataFrame()
     ):
         with caplog.at_level(logging.ERROR), pytest.raises(SystemExit) as exc:
-            importlib.import_module("scripts.fetch.fetch_yf_data").main()
+            importlib.import_module("gpt_trader.fetch.fetch_yf_data").main()
         assert exc.value.code == 1
         assert "No data available" in caplog.text
