+++ tests/test_parse_gpt_response.py
@@ -1,11 +1,8 @@
-import sys
 from pathlib import Path
 
 import pytest
 
-sys.path.insert(0, str(Path(__file__).resolve().parents[1]))
-
-from scripts.parse_response.parse_gpt_response import _extract_json
+from gpt_trader.parse.parse_gpt_response import _extract_json
 
 
 def test_extract_json_unfenced():
