+++ tests/test_run_step.py
@@ -4,9 +4,7 @@ from unittest.mock import patch
 
 import asyncio
 
-sys.path.insert(0, str(Path(__file__).resolve().parents[1]))
-
-from scripts.common import _run_step
+from gpt_trader.cli.common import _run_step
 
 
 class DummyProc:
@@ -23,12 +21,12 @@ def _fake_exec(*cmd):
 
 
 def test_run_step_py_module():
-    script = Path("scripts/fetch/fetch_mt5_data.py")
+    script = Path("src/gpt_trader/fetch/fetch_mt5_data.py")
     with patch("asyncio.create_subprocess_exec", side_effect=_fake_exec):
         asyncio.run(_run_step("fetch", script, "--foo"))
     assert _fake_exec.called[0] == sys.executable
     assert _fake_exec.called[1] == "-m"
-    assert _fake_exec.called[2].endswith("scripts.fetch.fetch_mt5_data")
+    assert _fake_exec.called[2].endswith("gpt_trader.fetch.fetch_mt5_data")
     assert _fake_exec.called[3:] == ("--foo",)
